chore(lint): enable pedantic warnings and refactor hotspots
