+++ test/unit/utils.test.ts
@@ -49,7 +49,9 @@ describe('resolveBase', () => {
 			path.join('C:/Users/user/AppData/Local', 'degit'),
 		);
 	});
+});
 
+describe('stashFiles and unstashFiles', () => {
 	it('stashes and unstashes nested directories with degit.json excluded from restore', () => {
 		const root = fs.mkdtempSync(path.join(process.cwd(), 'stash-'));
 		const cacheDir = path.join(root, 'cache');
@@ -86,7 +88,9 @@ describe('resolveBase', () => {
 			fs.rmSync(root, { force: true, recursive: true });
 		}
 	});
+});
 
+describe('fetch', () => {
 	it('resumes redirect responses when following a redirected archive fetch', async () => {
 		const createWriteStreamSpy = vi.spyOn(fs, 'createWriteStream').mockReturnValue({
 			on(event: string, handler: () => void) {
